+++ src/rp2040/gpio.c
@@ -274,7 +274,7 @@ bool serialise_gpio_config(const uint8_t gpio, struct NWBuffer* tx_buf) {
   reply.index = config.gpio[gpio].index;
   reply.address = config.gpio[gpio].address;
 
-  uint16_t tx_buf_len = pack_nw_buff(tx_buf, &reply, sizeof(reply));
+  size_t tx_buf_len = pack_nw_buff(tx_buf, &reply, sizeof(reply));
 
   if(!tx_buf_len) {
     printf("WARN: TX length greater than buffer size.\n");
